pyproject.toml
[project]
name = "predictsheet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-auth>=2.40.1",
    "gspread>=6.2.1",
    "matplotlib>=3.10.3",
    "numpy>=2.2.6",
    "oauth2client>=4.1.3",
    "pandas>=2.2.3",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "streamlit>=1.45.1",
]
